[doc] Remove unneded comment about collection refs
This concern is addressed by the following mail pullreq, that will be
into 0.4.0 release:

https://github.com/leapcode/leap_mail/pull/175

By the way, it is really funny that I cannot tell if this comment was
written by meskio, as the blame says, or if I was the original author of
the remark. I should stop drinking so much coffee.