[bug] fix concatenation of full_user_id to make mail work
At some moment, the semantics of the username that we pick from the
login widget changed, so that there's a duplication of the domain part
in the full_user_id that we were passing to the keymanager and the mail
services.

This fix has already been done in the current develop branch (for next
release 0.9.0), so there's no need to port it when merging it back.

While it is true that there's no official release of the client supporting
mail yet, this was an easy fix to port so we will do a 0.8.2
bugfix release so that the 0.8 branch can be used experimentally to test
against development providers. This doesn't mean that you can expect
support on the 0.8.x branch yet, it's only that there is one less bug in
here :)

Resolves: #6759
Releases: 0.8.2